requirements.txt
fastapi
uvicorn
jinja2
python-multipart
torch
torchvision
torchaudio
opencv-python
numpy
pandas
mediapipe
librosa
soundfile
moviepy
matplotlib
reportlab
pyyaml
pillow
scikit-image
